requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
confection==0.0.4
contexttimer==0.3.3
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
decorator==4.4.2
decord @ file:///Users/yogurt/nudge-lab/LLM/blip2/decord-0.6.0-cp311-none-macosx_10_15_x86_64.whl#sha256=a0eb1258beade34dceb29d97856a7764d179db1b5182899b61874f3418a1abc8
deepgram-sdk==0.0.0
distlib==0.3.6
einops==0.6.1
entrypoints==0.4
executing==1.2.0
fairscale==0.4.4
ffmpeg-python==0.2.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
huggingface-hub==0.12.1
identify==2.5.24
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
install==1.3.5
iopath==0.1.10
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kaggle==1.5.13
kiwisolver==1.4.4
langchain==0.0.177
langcodes==3.3.0
lazy_loader==0.2
llama-index==0.6.9
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.2.1
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.0rc1
nodeenv==1.7.0
numexpr==2.8.4
numpy==1.24.2
omegaconf==2.3.0
openai==0.27.6
openai-whisper==20230124
openapi-schema-pydantic==1.2.4
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.0
pandas==2.0.1
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.5.1
plotly==5.14.1
portalocker==2.7.0
pre-commit==3.3.1
preshed==3.0.8
proglog==0.1.10
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==13.3.5
rsa==4.9
salesforce-lavis==1.0.2
scikit-image==0.20.0
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.15
srsly==2.4.6
stack-data==0.6.2
streamlit==1.22.0
sympy==1.11.1
tenacity==8.2.2
text-unidecode==1.3
thinc==8.1.10
tifffile==2023.4.12
tiktoken==0.4.0
timm==0.4.12
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.0.dev20230305
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.13
validators==0.20.0
virtualenv==20.23.0
wasabi==1.1.1
wcwidth==0.2.6
webdataset==0.2.48
websockets==10.4
yarl==1.8.2
zipp==3.15.0
